Move `pkg/kubadm` to `cmd/kubeadm/app`, remove `cmd/manual.go`
