+++ cmd/kubeadm/app/master/kubeconfig.go
@@ -23,8 +23,8 @@ import (
 
 	// TODO: "k8s.io/client-go/client/tools/clientcmd/api"
 	clientcmdapi "k8s.io/kubernetes/pkg/client/unversioned/clientcmd/api"
-	kubeadmapi "k8s.io/kubernetes/pkg/kubeadm/api"
-	kubeadmutil "k8s.io/kubernetes/pkg/kubeadm/util"
+	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/api"
+	kubeadmutil "k8s.io/kubernetes/cmd/kubeadm/app/util"
 	certutil "k8s.io/kubernetes/pkg/util/cert"
 )
