--- pkg/kubeadm/.gitignore
@@ -1,1 +1,0 @@
-kubeadm
